% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{aggregate_monthly_to_annual}
\alias{aggregate_monthly_to_annual}
\title{Aggregate Monthly Data to Annual Totals}
\usage{
aggregate_monthly_to_annual(df)
}
\description{
Aggregate Monthly Data to Annual Totals
}
