% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{calculate_abundance}
\alias{calculate_abundance}
\title{Convert Nest Counts to Abundance (Nesters & Females)}
\usage{
calculate_abundance(df, clutch_freq = 5.5, remig_int = 3.06, quiet = FALSE)
}
\arguments{
\item{df}{Dataframe from prep_nesting_data (Year, Site, Count)}

\item{clutch_freq}{Nests per female per season}

\item{remig_int}{Average years between nesting seasons}
}
\description{
Convert Nest Counts to Abundance (Nesters & Females)
}
