% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_draws.R
\name{get_posteriors}
\alias{get_posteriors}
\title{Extract Posteriors and Calculate Final Abundance (Technical Memo Standard)}
\usage{
get_posteriors(res, remig_int = 3)
}
\description{
Extract Posteriors and Calculate Final Abundance (Technical Memo Standard)
}
