% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_turtle_status}
\alias{plot_turtle_status}
\title{Plot Turtle Population Status (Visualizing Imputed Trend)}
\usage{
plot_turtle_status(res, abund, title = "Population Trend")
}
\description{
Plot Turtle Population Status (Visualizing Imputed Trend)
}
