% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marss_engine.R
\name{run_marss_models}
\alias{run_marss_models}
\title{Run Comparative MARSS Models (Shared and Independent)}
\usage{
run_marss_models(df)
}
\arguments{
\item{df}{Cleaned abundance data frame from calculate_abundance()}
}
\value{
A list containing both the shared-trend and independent-trend MARSS fit objects
}
\description{
Run Comparative MARSS Models (Shared and Independent)
}
