% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{simulate_turtle_data}
\alias{simulate_turtle_data}
\title{Simulate Sea Turtle Nesting Data}
\usage{
simulate_turtle_data(
  n_years = 25,
  n_sites = 3,
  true_trend = -0.03,
  process_error = 0.1,
  nest_to_female = 5.5
)
}
\arguments{
\item{n_years}{Number of years to simulate}

\item{n_sites}{Number of nesting beaches}

\item{true_trend}{The annual growth rate (e.g., -0.03 is 3\% decline)}
}
\description{
Simulate Sea Turtle Nesting Data
}
