% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{calculate_abundance}
\alias{calculate_abundance}
\title{Convert Nest Counts to Abundance}
\usage{
calculate_abundance(df, clutch_freq = NULL, remig_int = NULL)
}
\arguments{
\item{df}{Cleaned dataframe from prep_nesting_data}

\item{clutch_freq}{Average nests per female per season}

\item{remig_int}{Average years between nesting seasons}
}
\description{
Convert Nest Counts to Abundance
}
