% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_turtle_status}
\alias{plot_turtle_status}
\title{Plot Population Trends and Status}
\usage{
plot_turtle_status(model_output, abundance_df, species_name = "Sea Turtle")
}
\arguments{
\item{model_output}{Result from run_turtle_model}

\item{abundance_df}{Result from calculate_abundance}

\item{species_name}{String for plot title}
}
\description{
Plot Population Trends and Status
}
