% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_turtle_status}
\alias{plot_turtle_status}
\title{Plot Turtle Population Status (Exact singleUQ Aesthetics)}
\usage{
plot_turtle_status(res, abund)
}
\description{
Plot Turtle Population Status (Exact singleUQ Aesthetics)
}
