% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prep_nesting_data}
\alias{prep_nesting_data}
\title{QAQC and Prep Nesting Data}
\usage{
prep_nesting_data(df)
}
\arguments{
\item{df}{A dataframe with Year, Site, and Count columns}
}
\description{
QAQC and Prep Nesting Data
}
