% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{run_turtle_model}
\alias{run_turtle_model}
\title{Run Bayesian Trend Analysis (Stabilized for Imputation)}
\usage{
run_turtle_model(df, iter = 1e+05, parallel = FALSE)
}
\description{
Run Bayesian Trend Analysis (Stabilized for Imputation)
}
