% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{run_turtle_model}
\alias{run_turtle_model}
\title{Run Bayesian Trend Analysis}
\usage{
run_turtle_model(df, iter = 1000, parallel = FALSE)
}
\arguments{
\item{df}{Dataframe from calculate_abundance}

\item{iter}{Number of MCMC iterations}

\item{parallel}{Logical. Run on multiple cores?}
}
\description{
Run Bayesian Trend Analysis
}
